Clean up directory/component structure
