+++ packages/frontend/web/components/Header/MainMenuToggle/VeggieBurger.tsx
